Owner view

+++ Art-Gallery/ViewModels/InventoryViewModel.cs
@@ -8,11 +8,11 @@ namespace Art_Gallery.ViewModels
     public class InventoryViewModel
     {
         public string Title { get; set; }
-        public float Cost { get; set; }
-        public float AskingPrice { get; set; }
+        public double Cost { get; set; }
+        public double AskingPrice { get; set; }
         public List<PieceViewModel> Pieces { get; set; }
         public bool Sold { get; set; }
-        public float TotalSales { get; set; }
-        public float TotalProfit { get; set; }
+        public double TotalSales { get; set; }
+        public double TotalProfit { get; set; }
     }
 }
